requirements.txt
torch>=1.12
torchvision
transformers
ultralytics    # для YOLOv8
timm
albumentations
pandas
numpy
scikit-learn
pyyaml
omegaconf
hydra-core
wandb
opencv-python
Pillow